modulo_04_langgraph_gcp/pyproject.toml
[project]
name = "modulo-04-langgraph-gcp"
version = "0.1.0"
description = "Módulo 04: Multi-Agent System con LangGraph y Vertex AI"
requires-python = ">=3.11"
dependencies = [
    # ── Google Generative AI (Gemini / Vertex AI unified SDK) ──────────────────
    "google-genai>=1.15.0",
    "google-cloud-aiplatform>=1.97.0",

    # ── LangChain core + integraciones Google ─────────────────────────────────
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-vertexai>=2.0.0",   # Mantener para compatibilidad ChromaDB
    "langchain-google-genai>=2.0.0",      # Clases actualizadas: ChatGoogleGenerativeAI
    "langchain-community>=0.3.0",

    # ── LangGraph para sistemas multi-agente ───────────────────────────────────
    "langgraph>=0.2.0",

    # ── RAG: vector stores + embeddings ───────────────────────────────────────
    "chromadb>=0.5.0",
    "langchain-chroma>=0.1.0",

    # ── Utilidades ─────────────────────────────────────────────────────────────
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "ipykernel>=6.0.0",
    "jupyterlab",
    "ruff>=0.4.0",
]
